adds demo recording tools
